+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [2.1.147] - 2026-07-15
+
+### Bug fixes
+
+- add actions:read to generated cd.yml release permissions (#2399)
+- reflow generated README description under markdownlint MD013 (#2400)
+
+### Features
+
+- run the finalize/release/install cascade on the relay branch path (#2401)
+
 ## [2.1.146] - 2026-07-15
 
 ### Features